requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.4.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
cryptography==46.0.3
cssselect2==0.8.0
Django==5.2.8
django-background-tasks==1.2.8
django-cors-headers==4.7.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
freetype-py==2.5.1
frozenlist==1.8.0
html5lib==1.1
idna==3.10
lxml==6.0.2
multidict==6.7.0
mysqlclient==2.2.7
num2words==0.5.14
oscrypto==1.3.0
pillow==11.1.0
propcache==0.4.1
pycairo==1.29.0
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.9.0
pypdf==6.3.0
python-bidi==0.6.7
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
razorpay==1.4.2
reportlab==4.4.5
requests==2.32.3
rlPyCairo==0.4.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
tinycss2==1.4.0
twilio==9.8.6
typing_extensions==4.12.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.3.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
yarl==1.22.0
